requirements.txt
# python requirements
pip>=9.0
absl-py==0.7.1
chumpy==0.68
deepdish==0.3.6
ipdb==0.12
matplotlib==3.0.3
numpy==1.16.3
opencv-python==3.4.5.20
scikit-image==0.15.0
scipy==1.2.1
tensorflow==1.8
torch==0.4.0
torchvision==0.2.2
tqdm==4.19.9
